requirements.txt
# fastapi-service/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
pydantic-settings
httpx
python-jose[cryptography]
passlib[bcrypt]
redis
aiogram